Try to make font keyword behaviour work better with mode hooks.
Currently, due to maybe legacy issues? The keyword table is built before
the mode hooks are run, but cperl variables change how the table is built.
The result of this is that you can't configure cperl mode in the cperl mode
hook, but need to drop the customisations into the global space.
This is an attempt to make this better.